requirements.txt
discord.py==2.3.1
psycopg2==2.9.3
